--- backend/src/main/java/com/is442g1t4/gpa/auth/RegisterRequest.java
@@ -1,17 +1,0 @@
-package com.is442g1t4.gpa.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-    private String name;
-    private String username;
-    private String password;
-    private String email;
-}
