+++ backend/src/main/java/com/is442g1t4/gpa/auth/email/EmailService.java
@@ -1,4 +1,4 @@
-package com.is442g1t4.gpa.auth;
+package com.is442g1t4.gpa.auth.email;
 
 import java.util.Map;
