+++ backend/src/main/java/com/is442g1t4/gpa/auth/utility/PasswordRequest.java
@@ -1,0 +1,15 @@
+package com.is442g1t4.gpa.auth.utility;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class PasswordRequest {
+    private String newpassword;
+    private String oldpassword;
+}
