+++ backend/src/main/java/com/is442g1t4/gpa/user/UserController.java
@@ -15,7 +15,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.is442g1t4.gpa.auth.PasswordRequest;
+
+import com.is442g1t4.gpa.auth.utility.PasswordRequest;
 import com.is442g1t4.gpa.portfolio.Portfolio;
 
 @RestController
